HomeViewModelをインタフェースと実装に分離
